UUID support in server
UUID support:
Anywhere a playername is used as parameter, a UUID is now also accepted.
An attempt is first made to convert the parameter to a UUID and find the
matching player.
If this fails (invalid UUID or no player found with UUID), an attempt is
made to find a player with the exact name provided.
For the required UUID format, search the Java documentation for
UUID.fromString()