Added EXTENDED_PLAYER_DATA setting
The EXTENDED_PLAYER_DATA config switch was added. When set to its
default value of false, invoking a POST request will only send elaborate
data about the sender.
When set to true, elaborate snapshots of all the players are added. Note
that this may increase the POST request data size a lot.