Removed 'Type', added support for dual wield
Type ID:
In the past item types (like in the player inventory) where passed as a
numerical ID.
Since this is now deprecated, the type is now only passed as a string.
Search the Bukkit API for the Material enum to find all possible names.
For example, a player holding a flower pot used to get serialized as
having an item with 'Type' 140 and 'TypeName' FLOWER_POT. The 'TypeName'
is still FLOWER_POT, but 'Type' was removed.

JSON Player.CurrentItemID:
In Minecraft 1.9 dual wield was introduced. To support this feature a
change was made to the JSON data sent in a request.
In earlier versions a field 'CurrentItemID' was included, this has now
been replaced with 'MainHandItemID' and 'OffHandItemID'.